Improve room registration logic, update rooms for Map.
Room registration has long been a slight issue, when concierge was non-
persistent we'd re-register in a loop in case concierge restarted.
When concierge was replaced by Map, which was persistent, we stopped
that looping.. but then if Room started before map did, we'd never
register.
This usually mean't we'd be ok, until we wiped the db down, and then had
to remember the right order to bring it all up in.. not good.

So this change adds the RoomRegistrationHandler that attempts
registration and tries to do the right thing based on the result of that
attempt. If not registered, the room is registered. If already
registered, the room is compared to see if an update is needed.

Additionally the exits obtained during registration are now remembered,
and used by the rooms as their exit info.

The exits command has been removed, because that's now handled by the
mediator.

Doors are added, to allow rooms to specify how their entrances should
look from other rooms.

Commands for a room are now sent as part of roomHello 'location' reply.